new pdf recipe

+++ src/command/render/cleanup.ts
@@ -33,7 +33,7 @@ export function cleanup(
 
   // if we aren't keeping the markdown and we are self-contained, then
   // delete the supporting files
-  if (!keepMd && selfContained) {
+  if (!keepMd && !format.keep?.tex && selfContained) {
     if (computations.supporting) {
       computations.supporting.forEach((path) => {
         Deno.removeSync(path, { recursive: true });
